requirements.txt
-i https://pypi.org/simple
amqp==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
argon2-cffi==20.1.0
asgiref==3.2.7; python_version >= '3.5'
billiard==3.6.3.0
bleach==3.1.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
celery==4.4.3
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==2.9.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django-celery-results==1.2.1
django-extensions==2.2.9
django==3.0.6
djangorestframework==3.11.0
docutils==0.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==2.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==1.6.0; python_version < '3.8'
jeepney==0.4.3; sys_platform == 'linux'
jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
joblib==0.15.1; python_version >= '3.6'
jwcrypto==0.7
keyring==21.2.1; python_version >= '3.6'
kombu==4.6.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
livereload==2.6.1
lunr[languages]==0.5.8
markdown==3.2.2; python_version >= '3.5'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
mkdocs==1.1.2
nltk==3.5
packaging==20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pkginfo==1.5.0.1
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygments==2.6.1; python_version >= '3.5'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2020.1
pyyaml==5.3.1
readme-renderer==26.0
regex==2020.5.14
requests-toolbelt==0.9.1
requests==2.23.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
secretstorage==3.1.2; sys_platform == 'linux'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
timeago==1.0.14
tornado==6.0.4; python_version >= '3.5'
tqdm==4.46.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
twine==3.1.1
urllib3==1.25.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
vine==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
webencodings==0.5.1
wheel==0.34.2
zipp==3.1.0; python_version >= '3.6'
